ARTHY-32 update picture on website

+++ App/Backend/Modules/Checklist/ChecklistController.php
@@ -28,9 +28,11 @@ class ChecklistController extends BackController
     public function executeShow(HTTPRequest $request)
     {
         $item = $this->managers->getManagerOf('Checklist')->getUnique($request->getData('id'));
-        $photo = $this->managers->getManagerOf('Photo')->getUnique($item->getIdPhoto());
+        if ($item->getIdPhoto() != -1) {
+            $photo = $this->managers->getManagerOf('Photo')->getUnique($item->getIdPhoto());
+            $this->page->addVar('photo', $photo);
+        }
         $this->page->addVar('item', $item);
-        $this->page->addVar('photo', $photo);
 
     }
 
@@ -43,57 +45,77 @@ class ChecklistController extends BackController
 
     public function processForm(HTTPRequest $request)
     {
+        $photoValide = true;
+
         $item = new Checklist([
             'titre' => $request->postData('titre'),
             'contenu' => $request->postData('contenu'),
             'isImportant' => ($request->postData('important')) ? 1 : 0
         ]);
 
-        $extUp = strtolower(substr(strrchr($_FILES['photo']['name'], '.'), 1));
-        $name = '/Picture/' . uniqid(rand(), true) . '.' . $extUp;
-        $dir = dirname(getcwd()) . "/Web" . $name;
-        $photo = new Photo([
-            'photo' => $name,
-            'chemin' => $dir
-        ]);
-        if (empty($_FILES['photo'])) {
-            $photo->setErreur(Photo::PICTURE_EMPTY);
-        }
-        if ($_FILES['photo']['error'] > 0) {
-            $photo->setErreur(Photo::INVALIDE);
-        }
-
-        if (in_array($extUp, Photo::EXTENSIONS)) {
-            $photo->setErreur(Photo::FORMAT);
-        }
-
-        if (empty($photo->erreurs())) {
-            $res = move_uploaded_file($_FILES['photo']['tmp_name'], $dir);
-            if (!$res) {
-                $photo->setErreur(Photo::INVALIDE);
-            }
-        }
-
-        if ($request->postExists('idPhoto')) {
-            $item->setIdPhoto($request->postData('idPhoto'));
-
-        }
-
         $alreadyIn = $this->managers->getManagerOf('Checklist')->alreadyIn($item);
 
         if ($request->postExists('id')) {
             $item->setId($request->postData('id'));
+            $item->setIdPhoto($request->postData('idPhoto'));
             if ($item->getTitre() == $request->postData('lastTitle')) {
                 $alreadyIn = false;
             }
         }
 
+        $empty = empty($_FILES['photo']['tmp_name']);
+
+        $photo = new Photo();
+
+        if (!$empty) {
+            $extUp = strtolower(substr(strrchr($_FILES['photo']['name'], '.'), 1));
+            $name = '/Picture/' . uniqid(rand(), true) . '.' . $extUp;
+            $dir = dirname(getcwd()) . "/Web" . $name;
+            $photo = new Photo([
+                'photo' => $name,
+                'chemin' => $dir,
+                'name' => $_FILES['photo']['name'],
+                'type' => $_FILES['photo']['type'],
+                'size' => $_FILES['photo']['size']
+            ]);
+            if (empty($_FILES['photo'])) {
+                $photo->setErreur(Photo::PICTURE_EMPTY);
+            }
+            if ($_FILES['photo']['error'] > 0) {
+                $photo->setErreur(Photo::INVALIDE);
+            }
+
+            if (in_array($extUp, Photo::EXTENSIONS)) {
+                $photo->setErreur(Photo::FORMAT);
+            }
+
+            if ($request->postExists('idPhoto')) {
+                $idPhoto = $request->postData('idPhoto');
+                $photo->setIdPhoto($idPhoto);
+                if ($idPhoto != -1) {
+                    $lastPicture = $this->managers->getManagerOf('Photo')->getUnique($item->getIdPhoto());
+                    if ($photo->equals($lastPicture)) {
+                        $photo = new Photo();
+                    }
+                }
+            }
+        }
+
+
         if ($alreadyIn) {
             $this->app->getUser()->setFlash('<span style="color: red"><strong>Cet item existe déja !</strong></span>');
-        } else if ($item->isValid() && !$alreadyIn && $photo->isValid()) {
-            $this->managers->getManagerOf('Checklist')->save($item, $photo);
-            $this->app->getUser()->setFlash($item->isNew() ? 'L\'item a bien été ajouté !' : 'L\'item a bien été modifié !');
-            $this->app->httpResponse()->redirect('/admin/checklist.html');
+        } else if ($item->isValid() && !$alreadyIn) {
+            if (!$photo->isEmpty()) {
+                if ($request->postExists('idPhoto') && $request->postData('idPhoto') != -1) {
+                    unlink($lastPicture->getChemin());
+                }
+                $photoValide = move_uploaded_file($_FILES['photo']['tmp_name'], $dir);
+            }
+            if ($photoValide) {
+                $this->managers->getManagerOf('Checklist')->save($item, $photo);
+                $this->app->getUser()->setFlash($item->isNew() ? 'L\'item a bien été ajouté !' : 'L\'item a bien été modifié !');
+                $this->app->httpResponse()->redirect('/admin/checklist.html');
+            }
         }
         $this->page->addVar('erreurs', $item->erreurs());
         $this->page->addVar('item', $item);
@@ -105,18 +127,22 @@ class ChecklistController extends BackController
             $this->processForm($request);
         }
         $item = $this->managers->getManagerOf('Checklist')->getUnique($request->getData('id'));
-        $photo = $this->managers->getManagerOf('Photo')->getUnique($item->getIdPhoto());
+        if ($item->getIdPhoto() != -1) {
+            $photo = $this->managers->getManagerOf('Photo')->getUnique($item->getIdPhoto());
+            $this->page->addVar('photo', $photo);
+        }
         $this->page->addVar('item', $item);
-        $this->page->addVar('photo', $photo);
     }
 
     public function executeDelete(HTTPRequest $request)
     {
         $item = $this->managers->getManagerOf('Checklist')->getUnique($request->getData('id'));
-        $photo = $this->managers->getManagerOf('Photo')->getUnique($item->getIdPhoto());
         $this->managers->getManagerOf('Checklist')->delete($request->getData('id'));
-        $this->managers->getManagerOf('Photo')->delete($photo->getIdPhoto());
-        unlink($photo->getChemin());
+        if ($item->getIdPhoto() != -1) {
+            $photo = $this->managers->getManagerOf('Photo')->getUnique($item->getIdPhoto());
+            $this->managers->getManagerOf('Photo')->delete($photo->getIdPhoto());
+            unlink($photo->getChemin());
+        }
 
         $this->app->getUser()->setFlash('L\'item a bien été supprimée !');
